+++ changelog.md
@@ -1,3 +1,22 @@
 # Changelog
 
-Snapshot 2021-11-24 15:29
+**./src/common/possible-criteria-validation.js**
+* Changes to 'loopCriteriaComplete'
+	* Removed parameters: 'givenDistance', 'givenStops'
+	* Declared new parameter 'givenRoute'
+	* Declared new variables: 'finalDistance', 'finalStops'
+	* Replaced 'givenStops' reference with 'finalStops'
+	* Replaced 'givenDistance' reference with 'finalDistance'
+* Wrote new functions:
+	* x
+	* x
+
+---
+
+**./src/possible-routes.js**
+* Changes to 'validateCompletedRoute'
+	* Removed parameters: 'distVal', 'stopCount'
+	* Declared new parameter 'compRoute'
+	* Updated 'possibleCriteriaValidation' call to use correct arguments.
+* Changes to 'iterateRoutes'
+	* Updated 'validateCompletedRoute' call to use correct arguments.